refactor: Unify names with handler prefixes.
