+++ src/server/middleware/authHandler.ts
