+++ src/server/middleware/extractTokenHandler.ts
